REFACTOR: Fase 2 — higiene de base antes da entrega
Limpeza de pré-produção sobre o main consolidado (Q.25+Q.26). Sem
alteração de comportamento — 2073 testes verdes, tsc/build OK.

ruff check src/: 115 -> 5 avisos. Inclui 3 bugs reais latentes:
- notify_listener: asyncio.create_task sem referência forte (RUF006) —
  a task podia ser recolhida pelo GC a meio, perdendo o push de regra.
- drift_detector: anotação "SheetData" sem import (F821) — corrigida
  com import sob TYPE_CHECKING.
- rag.py: import `text` da sqlalchemy morto — removido.
O resto são avisos cosméticos (loop vars/locais não usados). Restam 5
ASYNC240 (operações de ficheiro pontuais em endpoints/scheduler) —
não-bug, deixados de propósito.

requirements.txt: adicionar aioodbc + pyodbc — o adapter read-only do
ERP MAR-KAYAKS usa-os; estavam no venv mas não no ficheiro (deploy não
reproduzível).

frontend: remover 13 console.log de debug do CopilotDrawer.

+++ src/ml/models_domain/duration.py
@@ -91,7 +91,7 @@ class DurationModel:
                 range(len(filtered)),
                 key=lambda i: timestamps[i],
             )
-            split_at = int(round(len(order_idx) * 0.8))
+            split_at = round(len(order_idx) * 0.8)
             train_idx = order_idx[:split_at]
             val_idx = order_idx[split_at:]
             filtered = [filtered[i] for i in train_idx + val_idx]
