+++ src/ml/models_domain/quality_risk.py
@@ -78,7 +78,7 @@ class QualityRiskModel:
 
         if has_temporal:
             order_idx = sorted(range(len(rows)), key=lambda i: timestamps[i])
-            split_at = int(round(len(order_idx) * 0.8))
+            split_at = round(len(order_idx) * 0.8)
             train_idx = order_idx[:split_at]
             val_idx = order_idx[split_at:]
             sorted_rows = [rows[i] for i in train_idx + val_idx]
